+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Added `SignatureAlgorithm.jcaSignature` to extract the JCA signature for a TLS signature algorithm.
 - Added `Signer` utility class to easily sign and verify with a private and public key.
 - Renamed `IO` class to `IOStreams` to not create confusion because Java 25 will add an `IO` class.
+- Changed `OAuth2ApiClient` constructors to use `JwsAlgorithm` instead of `String`.
 
 ---